pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sensorsim"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "scipy",
    "numba",
    "open3d==0.19.0",
    "opencv-python-headless",
    "plyfile",
    "PyYAML",
    "mcap-ros2-support",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
include = ["sim*"]